clip on code added | untested

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/subsystems/Arm.java
@@ -72,7 +72,7 @@ public class Arm {
 
     }
 
-    public void pitchAppend(float pos) {
+    public void pitchAppend(double pos) {
         pitchPos += pos;
         pitchSet(pitchPos);
     }
